Add google.Timestamp protobuf file
This is a temporary fix for the new feature flag being added to `demo.proto` we will probably replace how we generate golang files from the proto schema. Possibly use a protoc docker image or add a `go get google.golang.org/protobuf/types/known/timestamppb` but this wasn't as simple as adding that line to the `Dockerfile`'s in question.