feat: rewrite emailservice in ruby
Rather than use GRPC, we reimplement the emailservice as a simple HTTP
application, using Sinatra and OpenTelemetry auto-instrumentation. We
can do this because our use of GRPC in this instance is not very
complicated - we can just serialize the protobuf request to JSON and
send it over HTTP instead, and it's pretty straightforward.

Probably the most interesting part here is that we're rendering the
email text right into the logfile, which might be noisy. We can revisit
that decision if needed.