feat: modify setup to work with otel/ruby
Because we rewrote the emailservice as an HTTP service, we need to
modify the `EMAIL_SERVICE_ADDR` to reflect the actual scheme (`http`).
We could just do some more string concatenation in the app, I guess.

We pin to a specific collector version - I was getting weird segfaults
with whatever version of the collector was already on my machine, and
they were resolved with 0.52.0. However, that also required a minor
config change to the jaeger exporter. Yay, M1 macs!

We can't specify the otlp/grpc port for the ruby libraries, it won't
work. So for email service we pass the correct OTLP/HTTP URL in the env
var.