Rewrite emailservice in ruby (#109)
* feat: rewrite emailservice in ruby

Rather than use GRPC, we reimplement the emailservice as a simple HTTP
application, using Sinatra and OpenTelemetry auto-instrumentation. We
can do this because our use of GRPC in this instance is not very
complicated - we can just serialize the protobuf request to JSON and
send it over HTTP instead, and it's pretty straightforward.

Probably the most interesting part here is that we're rendering the
email text right into the logfile, which might be noisy. We can revisit
that decision if needed.

* feat: call emailservice via HTTP POST

This commit causes the checkoutservice to call emailservice via HTTP
POST. We leverage the golang net/http autoinstrumentation to create the
spans and propagate context correctly for us.

* feat: modify setup to work with otel/ruby

Because we rewrote the emailservice as an HTTP service, we need to
modify the `EMAIL_SERVICE_ADDR` to reflect the actual scheme (`http`).
We could just do some more string concatenation in the app, I guess.

We pin to a specific collector version - I was getting weird segfaults
with whatever version of the collector was already on my machine, and
they were resolved with 0.52.0. However, that also required a minor
config change to the jaeger exporter. Yay, M1 macs!

We can't specify the otlp/grpc port for the ruby libraries, it won't
work. So for email service we pass the correct OTLP/HTTP URL in the env
var.

* docs: note that emailservice is now ruby

* fixup: remove `loglevel: debug` from otel-col config

Whoops, this was accidentally left in from my testing.

* fixup: appease markdownlint