ruby: switch to `-slim` image
If I had invested, say, a mere modicum of effort into understanding what
needed to be done to use a `-slim` image in the first place, then we
would have included this in #109 from the start. However, I clearly did not. :facepalm:

This change switches us to the `-slim` image, and installs the one thing
we actually need to build `puma`: Debian's `build-essential` package.