Update featureflagservice environment defaults
The feature flag service won't start unless certain environment
variables are set. For ease of set up, this provides default
values. In the case of ports, these defaults match those
in .env

Perhaps in a future update the helm installation could automatically
generate a secret to use for the `SECRET_KEY_BASE` variable.