shippingservice: Add http client request span
The commit adds new http client span for outgoing calls to `/getquote`
endpoint of quoteservice via the `reqwests` client middleware instrumentation.
This provides a bit of the server/client tasks separation visibility
as well as enables span pairing with the downstream quoteservice server span.
The instrumentation uses the available `reqwests-tracing` and `tracing-opentelementry` crates.