feat(load-generator): Adapt OpenFeature OFREP instead of IN_PROCESS
OpenFeature supports multiple different modes.
The newest implementation relies on threads and other python internal,
which seem to be monkeypatched by locust.
Due to this, the IN_PROCESS mode is not working properly anymore and
is causing issues.

Instead we suggest to switch to the OFREP (OpenFeature Remote Evaluation Protocol)
for this usecase. This is a simple http call which will return all evaluate flags,
and will regularly poll for changes.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>